charaobj: type setSta 'current' as int for signed-compare match
The shipped binary compares the status slot against 0 with cmpwi
(signed) after the lhz load. Widening 'current' to int matches that
form. Comparison is == 0, so semantics are unchanged.
setSta 67.65% -> 67.72%.

+++ src/charaobj.cpp
@@ -1860,7 +1860,7 @@ void CGCharaObj::setSta(int staIndex, int value)
 	}
 
 	unsigned char* staSlot = reinterpret_cast<unsigned char*>(m_scriptHandle) + (staIndex * 2);
-	unsigned short current = *reinterpret_cast<unsigned short*>(staSlot + 0x3E);
+	int current = *reinterpret_cast<unsigned short*>(staSlot + 0x3E);
 	value &= ~(value >> 31);
 
 	if (current == 0 && value != 0) {
